Modifying how gl functions are loaded
Wrapping all gl functions in a class and creating an instance of that
class to access them. This class has a method to load all the functions,
which means it will load all functions used and should work cross
platform.  This presents a tidy interface for accessing gl functions and
as stated should work cross platform.